# Reduce allocations in vectorized operations