--- a/src/ACSetInterface.jl
+++ b/src/ACSetInterface.jl
@@ -5,7 +5,8 @@ export ACSet, acset_schema, acset_name, dom_parts, codom_parts, subpart_type,
   rem_part!, rem_parts!, cascading_rem_part!, cascading_rem_parts!, gc!,
   copy_parts!, copy_parts_only!, disjoint_union, tables, pretty_tables,
   @acset, constructor, undefined_subparts, PartsType, DenseParts, MarkAsDeleted,
-  rem_free_vars!, parts_type
+  rem_free_vars!, parts_type,
+  ACSetAllocTest
 
 using MLStyle: @match
 using StaticArrays: StaticArray
@@ -267,6 +268,13 @@ function set_subpart! end
   end
 end
 
+@inline function set_subpart!(acs::ACSet , parts::Union{AbstractVector{Int}, AbstractSet{Int}}, ::Type{Val{name}}, vals) where name
+  broadcast(parts, vals) do part, val
+    set_subpart!(acs, part, name, val)
+  end
+end
+
+
 """ Mutate subparts of a part in a C-set.
 
 Both single and vectorized assignment are supported.
@@ -278,6 +286,9 @@ See also: [`set_subpart!`](@ref).
 end
 
 @inline set_subparts!(acs, part; kw...) = set_subparts!(acs, part, (;kw...))
+
+struct ACSetAllocTest end
+@inline Base.setindex!(acs::ACSet, val, part, name, ::ACSetAllocTest) = set_subpart!(acs, part, Val{name}, val)
 
 @inline Base.setindex!(acs::ACSet, val, part, name) = set_subpart!(acs, part, name, val)
 @inline Base.setindex!(acs::ACSet, vals, name) = set_subpart!(acs, name, vals)