This PR is meant as a quick fix for issue #135. It is a non-breaking fix as I've run the tests locally and all have passed. As of writing this I have not yet looked at vectorized gets or using the colon and these might still be issues. 

New benchmarks from initial example: https://github.com/AlgebraicJulia/ACSets.jl/issues/135#issue-2306532997
```julia
[ Info: Using add_vector
  0.000010 seconds (1 allocation: 144 bytes)

[ Info: Using add_loop
  0.000006 seconds
```

So using the vectorized setting now seems to be much improved. There is the issue that there are still some allocations but this is much improved.

Benchmarks run mirroring this later example: https://github.com/AlgebraicJulia/ACSets.jl/issues/135#issuecomment-2123049737
```julia
using ACSets

SchTest = BasicSchema([:V,:E], [(:v0,:E,:V)])

@acset_type Test(SchTest, index=[:v0])

s = Test()

N = 100
add_parts!(s, :V, N)
idx = add_parts!(s, :E, N)

entry = ones(Int64, N);

function add_broadcast(s, idx, entry, symbol)
  broadcast(idx, entry) do i, id
    set_subpart!(s, i, symbol, id)
  end
end

function add_vector(s, idx, symbol, entry)
  s[idx, symbol] = entry
end

function add_subpart_reg(s, idx, symbol, entry)
  set_subpart!(s, idx, symbol, entry)
end

function add_subpart_val(s, idx, symbol, entry)
  set_subpart!(s, idx, Val{symbol}, entry)
end

begin
  add_broadcast(s, idx, entry, :v0)
  add_vector(s, idx, :v0, entry)
  add_subpart_reg(s, idx, :v0, entry)
  add_subpart_val(s, idx, :v0, entry)
end;

@info "Using add_broadcast"
@time add_broadcast(s, idx, entry, :v0);

@info "Using add_vector"
@time add_vector(s, idx, :v0, entry);

@info "Using add_subpart_reg"
@time add_subpart_reg(s, idx, :v0, entry);

@info "Using add_subpart_val"
@time add_subpart_val(s, idx, :v0, entry);

```

```julia
[ Info: Using add_broadcast
  0.000055 seconds (108 allocations: 7.547 KiB)

[ Info: Using add_vector
  0.000021 seconds (4 allocations: 1.031 KiB)

[ Info: Using add_subpart_reg
  0.000121 seconds (4 allocations: 1.031 KiB)

[ Info: Using add_subpart_val
  0.000022 seconds (4 allocations: 1.031 KiB)
```
These show that using either ```set_subparts!``` or the regular setting operator carry the benefits while the old version of broadcast is used to show the original allocations.