@GeorgeR227 2024-05-22
Quick benchmarks I took comparing dual mesh creating from CombinatorialSpaces.jl.

Old:
<img width="496" alt="image" src="https://github.com/AlgebraicJulia/ACSets.jl/assets/78235421/1e2ae8e6-0c5b-4478-abb4-5c9c668a85be">

New:
<img width="515" alt="0 036080 seconds (13 44 k allocations 1 755 MiB, 94 02 gc time)" src="https://github.com/AlgebraicJulia/ACSets.jl/assets/78235421/0818af83-5078-4e9c-aa99-184717034c0b">

Note that while ```subdivide_duals!``` is unchanged, the initial dual mesh construction is significantly improved.