@epatters 2024-05-22
Thanks George! I was going to request that you add a test but I see you've just done that.

When you're finished, mark the PR as ready for review and I'll merge it.